% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bps_annual_by_county}
\alias{bps_annual_by_county}
\title{Census Building Permit Survey data by county}
\format{
A data frame with lots variables:
\describe{
  \item{year}{year of survey}
  \item{fips}{county fips}
  \item{name}{county name}
  \item{total_units}{estimated total units}
  \item{units_1}{estimated 1-unit units}
  \item{units_2}{estimated 2-unit units}
  \item{units_3_4}{estimated 3-4 unit units}
  \item{units_5_more}{estimated 5+ unit units}
  \item{units_rep_1}{reported 1-unit units}
  \item{units_rep_2}{reported 2-unit units}
  \item{units_rep_3_4}{reported 3-4 unit units}
  \item{units_rep_5_more}{reported 5+ unit units}
  \item{buildings_1}{estimated 1-unit buildings}
  \item{buildings_2}{estimated 2-unit buildings}
  \item{buildings_3_4}{estimated 3-4 unit buildings}
  \item{buildings_5_more}{estimated 5+ unit buildings}
  \item{buildings_rep_1}{reported 1-unit buildings}
  \item{buildings_rep_2}{reported 2-unit buildings}
  \item{buildings_rep_3_4}{reported 3-4 unit buildings}
  \item{buildings_rep_5_more}{reported 5+ unit buildings}
  \item{value_1}{estimated 1-unit value}
  \item{value_2}{estimated 2-unit value}
  \item{value_3_4}{estimated 3-4 unit value}
  \item{value_5_more}{estimated 5+ unit value}
  \item{value_rep_1}{reported 1-unit value}
  \item{value_rep_2}{reported 2-unit value}
  \item{value_rep_3_4}{reported 3-4 unit value}
  \item{value_rep_5_more}{reported 5+ unit value}
  \item{survey_date}{survey date}
  \item{fips_state}{state fips}
  \item{fips_county}{county fips (without state included)}
  \item{region_code}{Census region code}
  \item{division_code}{Census division code}
}
}
\source{
\url{https://www.census.gov/construction/bps/}
}
\usage{
bps_annual_by_county
}
\description{
A dataset containing annual bps data by county, for 2000-2021.
}
\keyword{datasets}
