% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bps_annual_by_msa}
\alias{bps_annual_by_msa}
\title{Census Building Permit Survey data by msa}
\format{
A data frame with the following variables:
\describe{
  \item{year}{year of survey}
  \item{name}{msa name}
  \item{total_units}{total units}
  \item{units_1}{1-unit units}
  \item{units_2}{2-unit units}
  \item{units_3_4}{3-4 unit units}
  \item{units_5_more}{5+ unit units}
  \item{structures_units_5_more}{5+ unit structures}
  \item{note}{note}
}
}
\source{
\url{https://www.census.gov/construction/bps/}
}
\usage{
bps_annual_by_msa
}
\description{
A dataset containing annual bps data by msa, for 2004-2014.
}
\keyword{datasets}
