% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_msa.R
\name{by_msa}
\alias{by_msa}
\title{Get BPS data by metropolitan statistical area}
\usage{
by_msa(year)
}
\arguments{
\item{year}{Survey year to get}
}
\description{
Get BPS data by metropolitan statistical area
}
